@@ -1,38 +1,33 @@
 source 'https://rubygems.org'
 
-ruby '2.1.2'
+#ruby '2.1.2'
 
-gem 'rails', '3.2.6'
-gem 'devise', '2.1.0'
+gem 'rails', '4.2.3'
+gem 'devise'
 gem 'kaminari'
-gem 'gravatar_image_tag'
 gem 'rails_admin'
 gem 'fastercsv'
 gem 'cancan'
-gem 'twitter', '4.3.0'
-#gem 'rmagick'#, :require => 'RMagick'
-gem 'gruff'
 gem 'elo'
 gem 'execjs'
-gem 'therubyracer', '0.12.0'
+gem 'therubyracer'
 gem 'sqlite3'
 gem 'json'
+gem 'bootstrap-sass'
+gem 'gravatar'
+gem 'gravatar_image_tag'
 
 # Gems used only for assets and not required
 # in production environments by default.
-group :assets do
  # gem 'less-rails'
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+gem 'sass-rails'
+gem 'coffee-rails'
  # gem 'twitter-bootstrap-rails', '~> 2.1.9'
-
-
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   # gem 'therubyracer', :platforms => :ruby
 
-  gem 'uglifier', '>= 1.0.3'
-end
+#  gem 'uglifier', '>= 1.0.3'
 
 gem 'jquery-rails'
 
@@ -52,3 +47,5 @@ gem 'capistrano', '3.0.1'
 
 # To use debugger
 # gem 'ruby-debug'
+gem "rspec-rails", :group => [:development, :test]
+gem "factory_girl_rails", :group => [:development, :test]
