@@ -2,7 +2,7 @@ class Blog < ActiveRecord::Base
 
   belongs_to :player
 #  before_filter :authenticate_player!
-  attr_accessible :author, :contents, :summary, :title
+ # attr_accessible :author, :contents, :summary, :title
 
 
 
