@@ -11,6 +11,6 @@ class Player < ActiveRecord::Base
          :recoverable, :rememberable, :trackable, :validatable
 
   # Setup accessible (or protected) attributes for your model
-  attr_accessible :email, :name, :phone, :twitter, :password, :password_confirmation, :remember_me, :admin, :active
+  #attr_accessible :email, :name, :phone, :twitter, :password, :password_confirmation, :remember_me, :admin, :active
   # attr_accessible :title, :body
 end
