@@ -1,3 +1,3 @@
 class Season < ActiveRecord::Base
-  attr_accessible :end_date, :start_date
+ # attr_accessible :end_date, :start_date
 end
