@@ -15,7 +15,7 @@ set :stage, :production
 # definition into the server list. The second argument
 # something that quacks like a hash can be used to set
 # extended properties on the server.
-server '188.226.225.77', user: 'root', roles: %w{web app} #, my_property: :my_value
+server '46.101.41.131', user: 'deploy', roles: %w{web app} #, my_property: :my_value
 
 # you can set custom ssh options
 # it's possible to pass any option but you need to keep in mind that net/ssh understand limited list of options
