@@ -1,8 +1,8 @@
-require 'twitter'
-
-Twitter.configure do |config|
-   config.consumer_key = "MmLpCfZryJpziDQrP6v2fA"
-   config.consumer_secret = "r1JnVOv0fqpKWf85PYy7NqIeujLlso7Rz77dMBz0GJM"
-   config.oauth_token = "304956678-t1zBhgd9WPsLt2iPziMtMJUky7N67At8sBJOLVtE"
-   config.oauth_token_secret = "S1Y9hkH9Sx9HOvXHzFDpceX1JyNZjKveaWmUl0QaMQ"
-end
+#require 'twitter'
+#
+#Twitter.configure do |config|
+#   config.consumer_key = "MmLpCfZryJpziDQrP6v2fA"
+#   config.consumer_secret = "r1JnVOv0fqpKWf85PYy7NqIeujLlso7Rz77dMBz0GJM"
+#   config.oauth_token = "304956678-t1zBhgd9WPsLt2iPziMtMJUky7N67At8sBJOLVtE"
+#   config.oauth_token_secret = "S1Y9hkH9Sx9HOvXHzFDpceX1JyNZjKveaWmUl0QaMQ"
+#end
